backend/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-CORS==5.0.0
PyJWT==2.10.1
python-dotenv==1.0.1
Werkzeug==3.1.3
bcrypt==4.2.1
